requirements.txt
Flask
flask-cors
torch
torchvision
Pillow
requests
openai
Werkzeug
google-cloud-storage
gunicorn
